Updating my old fork

+++ build.fsx
@@ -10,7 +10,7 @@ open System.Xml.Linq
 open Fake
 open Fake.FileUtils
 
-let version = "0.4.0.1001"
+let version = "0.4.0.2002"
 let buildDir = "merged"
 let nugetDir = "nuget"
 let nugetDocs = nugetDir @@ "content"
@@ -35,48 +35,38 @@ let sampleSln = slnBuild "SampleSolrApp.sln"
 
 let nuGetBuild = Nu.build version
 
-Target "Clean" (fun _ -> 
+Target "Clean" <| fun _ -> 
     mainSln "Clean"
     sampleSln "Clean"
     rm_rf buildDir
     rm_rf nugetDir
-)
 
-Target "Build" (fun _ -> mainSln "Rebuild")
-Target "BuildSample" (fun _ -> sampleSln "Rebuild")
+Target "Build" <| fun _ -> mainSln "Rebuild"
+Target "BuildSample" <| fun _ -> sampleSln "Rebuild"
 
-let libs = ["SolrNet"; "SolrNet.DSL"; "HttpWebAdapters"; "Castle.Facilities.SolrNetIntegration"; "Ninject.Integration.SolrNet"; "NHibernate.SolrNet"; "StructureMap.SolrNetIntegration"; "AutofacContrib.SolrNet"; "Unity.SolrNetIntegration"]
+let libs = ["SolrNet"; "SolrNet.DSL"; "HttpWebAdapters"; "Castle.Facilities.SolrNetIntegration"; "Ninject.Integration.SolrNet"; "StructureMap.SolrNetIntegration"; "AutofacContrib.SolrNet"; "Unity.SolrNetIntegration"; "NHibernate.SolrNet"]
 let dlls = [for l in libs -> l + ".dll"]
 let dirs = [for l in libs -> l @@ "bin" @@ config]
 
-let testAssemblies = !! ("**/bin/"+config+"/*Tests.dll")
+let testAssemblies = !! ("**/bin/"+config+"/*Tests.dll") |> Seq.distinctBy (fun p -> p.Split [|'/';'\\'|] |> System.Linq.Enumerable.Last)
 let noIntegrationTests = "exclude Category: Integration"
 let onlyIntegrationTests = "Category: Integration"
+let testTargets = List.map (fun lib -> "Test." + lib) libs |> List.filter (fun l -> not (l.Contains "NHibernate"))
 
-Target "Test" (fun _ ->
-    testAssemblies |> Gallio.Run (fun p -> { p with Filters = noIntegrationTests })
-)
-
-for lib in libs do
-    Target ("Test." + lib) (fun _ ->
-        !! ("**/bin/"+config+"/"+lib+".Tests.dll")
+for lib,target in Seq.zip libs testTargets do
+    Target target <| fun _ ->
+        !! (lib+".Tests/bin/"+config+"/"+lib+".Tests.dll")
             |> Gallio.Run (fun p -> { p with Filters = noIntegrationTests })
-    )
 
-Target "Coverage" (fun _ ->
+Target "Coverage" <| fun _ ->
     testAssemblies |> Gallio.Run (fun p -> { p with 
                                                 Filters = noIntegrationTests
                                                 RunnerType = "NCover"
                                                 PluginDirectories = ["lib"] })
-)
 
-Target "IntegrationTest" (fun _ ->
-    Solr.start()
-    try
-        testAssemblies |> Gallio.Run (fun p -> { p with Filters = onlyIntegrationTests })
-    finally
-        Solr.stop()
-)
+Target "IntegrationTest" <| fun _ ->
+    use s = Solr.start()
+    testAssemblies |> Gallio.Run (fun p -> { p with Filters = onlyIntegrationTests })
 
 let merge libraries = 
     rm_rf buildDir
@@ -93,19 +83,17 @@ let merge libraries =
                             XmlDocs = true
                        }) output main
 
-Target "Merge" (fun _ ->
+Target "Merge" <| fun _ ->
     dlls |> Seq.skip 1 |> merge
-)
 
-Target "BasicMerge" (fun _ ->
+Target "BasicMerge" <| fun _ ->
     dlls |> Seq.skip 1 |> Seq.take 2 |> merge
-)
 
-Target "Version" (fun _ ->
+Target "Version" <| fun _ ->
     for l in libs do
-        AssemblyInfo (fun p -> { p with
+        AssemblyInfo <| fun p -> { p with
                                     OutputFileName = l @@ "Properties\\AssemblyInfo.cs"
-                                    CLSCompliant = true
+                                    CLSCompliant = Some true
                                     AssemblyTitle = l
                                     AssemblyDescription = l
                                     AssemblyProduct = l
@@ -113,26 +101,25 @@ Target "Version" (fun _ ->
                                     AssemblyCopyright = "Copyright Mauricio Scheffer 2007-" + DateTime.Now.Year.ToString()
                                     Guid = "6688f9b4-5f2d-4fd6-aafc-3a81c84a69f1"
                                     AssemblyVersion = version
-                                    AssemblyFileVersion = version })
-)
+                                    AssemblyFileVersion = version }
 
-Target "Docs" (fun _ ->
+Target "Docs" <| fun _ ->
     rm_rf docsDir
     let r = Shell.Exec(@"tools\doxygen\doxygen.exe")
     if r <> 0 then failwith "Doxygen failed"
-    rm docsFile
-    Rename docsFile (docsDir @@ "html\\index.chm")
+    if File.Exists docsFile then
+        rm docsFile
+        Rename docsFile (docsDir @@ "html\\index.chm")
     rm_rf docsDir
-)
 
-Target "NuGet" (fun _ ->
+Target "NuGet" <| fun _ ->
     rm_rf nugetDir
     mkdir nugetDocs
     mkdir nugetLib
-    cp docsFile nugetDocs
+    if File.Exists docsFile then
+        cp docsFile nugetDocs
     !!(buildDir @@ "SolrNet.*") |> Copy nugetLib
-    nuGetBuild "SolrNet" "Apache Solr client" ["CommonServiceLocator", "1.0"]
-)
+    nuGetBuild "SolrNet" "Apache Solr client" ["CommonServiceLocator", "[1.0]"]
 
 let nuGetSingle dir =
     rm_rf nugetDir
@@ -140,59 +127,54 @@ let nuGetSingle dir =
     !!(dir @@ "bin" @@ config @@ (dir + ".*")) |> Copy nugetLib
     nuGetBuild 
 
-Target "NuGet.Windsor" (fun _ ->
+let solrNetDep = "SolrNet", "[" + version + "]"
+
+Target "NuGet.Windsor" <| fun _ ->
     nuGetSingle 
         "Castle.Facilities.SolrNetIntegration" 
         "SolrNet.Windsor"
         "Windsor facility for SolrNet"
-        ["Castle.Windsor", "[2.5.1,2.5.3]"; "SolrNet", version]
-)
+        ["Castle.Windsor", "3.2.0"; solrNetDep]
 
-Target "NuGet.Ninject" (fun _ ->
+Target "NuGet.Ninject" <| fun _ ->
     nuGetSingle 
         "Ninject.Integration.SolrNet" 
         "SolrNet.Ninject"
         "Ninject module for SolrNet"
-        ["Ninject", "[2.2.0.0,2.2.1.0]"; "SolrNet", version]
-)
+        ["Ninject", "3.0.1.10"; solrNetDep]
 
-Target "NuGet.NHibernate" (fun _ ->
+Target "NuGet.NHibernate" <| fun _ ->
     nuGetSingle 
         "NHibernate.SolrNet" 
         "SolrNet.NHibernate"
         "NHibernate integration for SolrNet"
-        ["NHibernate", "3.1.0.4000"; "CommonServiceLocator", "1.0"; "SolrNet", version]
-)
+        ["NHibernate", "[3.2.0.4000]"; solrNetDep]
 
-Target "NuGet.StructureMap" (fun _ ->
+Target "NuGet.StructureMap" <| fun _ ->
     nuGetSingle 
         "StructureMap.SolrNetIntegration" 
         "SolrNet.StructureMap"
         "StructureMap registry for SolrNet"
-        ["structuremap", "2.6.2.0"; "SolrNet", version]
-)
+        ["structuremap", "[2.6.2.0]"; solrNetDep]
 
-Target "NuGet.Autofac" (fun _ ->
+Target "NuGet.Autofac" <| fun _ ->
     nuGetSingle 
         "AutofacContrib.SolrNet" 
         "SolrNet.Autofac"
         "Autofac module for SolrNet"
-        ["Autofac", "2.2.4.900"; "SolrNet", version]
-)
+        ["Autofac", "[2.6.3.862]"; solrNetDep]
 
-Target "NuGet.Unity" (fun _ ->
+Target "NuGet.Unity" <| fun _ ->
     nuGetSingle 
         "Unity.SolrNetIntegration" 
         "SolrNet.Unity"
         "Unity integration for SolrNet"
-        ["Unity", "2.1.505.0"; "SolrNet", version]
-)
+        ["Unity", "[2.1.505.0]"; solrNetDep]
 
-Target "ReleasePackage" (fun _ -> 
+Target "ReleasePackage" <| fun _ -> 
     let outputPath = "build"
     rm_rf outputPath
     mkdir outputPath
-    cp "readme.txt" outputPath
     if File.Exists docsFile then
         cp docsFile outputPath
 
@@ -218,9 +200,8 @@ Target "ReleasePackage" (fun _ ->
     !! (outputPath @@ "**\\*") |> Zip outputPath zipFile
 
     rm_rf outputPath
-)
 
-Target "PackageSampleApp" (fun _ ->
+Target "PackageSampleApp" <| fun _ ->
     let outputSolr = buildDir @@ solr
     cp_r solr outputSolr
     rm_rf (outputSolr @@ "solr\\data")
@@ -258,7 +239,6 @@ Target "PackageSampleApp" (fun _ ->
     
     !! (buildDir @@ "**\\*")
         |> Zip buildDir ("SolrNet-"+version+"-sample.zip")
-)
 
 
 Target "BuildAll" DoNothing
@@ -266,13 +246,16 @@ Target "TestAndRelease" DoNothing
 Target "BuildAndRelease" DoNothing
 Target "NuGet.All" DoNothing
 Target "All" DoNothing
+Target "Test" DoNothing
+Target "TestAndNuGet" DoNothing
 
-"Test" <== ["BuildAll"]
+"Test" <== ["BuildAll"] @ testTargets
 "BuildAll" <== ["Build";"Merge";"BuildSample"]
 "BuildAndRelease" <== ["Clean";"Version";"BuildAll";"Docs";"ReleasePackage"]
 "TestAndRelease" <== ["Clean";"Version";"Test";"ReleasePackage"]
 "NuGet" <== ["Clean";"Build";"BasicMerge";"Docs"]
 "NuGet.All" <== (getAllTargetsNames() |> List.filter ((<*) "NuGet") |> List.filter ((<>) "NuGet.All") |> List.sort)
+"TestAndNuGet" <== ["Clean";"Version"; "Test"; "NuGet.All"]
 "All" <== ["BuildAndRelease";"PackageSampleApp";"NuGet.All"]
 
 Run target
